refactor: hook, type 분리 #12
hook 파일에 정의되어있던 type 코드 분리
